Remove Gradle and build directories from Git tracking
