--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Thu Jan 16 16:35:53 KST 2025
-gradle.version=8.11.1
