+++ src/main/java/com/example/backend/document/repository/DocumentRepository.java
@@ -1,0 +1,8 @@
+package com.example.backend.document.repository;
+
+import com.example.backend.document.entity.Document;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DocumentRepository extends JpaRepository<Document, Long> {
+}
+
