+++ src/main/java/com/example/backend/member/DTO/MemberDTO.java
@@ -1,0 +1,4 @@
+package walab.cap2.dongeui.member;
+
+public class MemberDTO {
+}
