+++ src/main/java/com/example/backend/member/controller/MemberController.java
@@ -1,0 +1,4 @@
+package walab.cap2.dongeui.member;
+
+public class MemberController {
+}
