+++ src/main/java/com/example/backend/member/repository/MemberRepository.java
@@ -1,0 +1,6 @@
+package walab.cap2.dongeui.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
