--- src/main/java/com/example/backend/repository/DocumentRepository.java
@@ -1,8 +1,0 @@
-package com.example.backend.repository;
-
-import com.example.backend.entity.Document;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DocumentRepository extends JpaRepository<Document, Long> {
-}
-
