+++ src/main/java/com/example/backend/signature/DTO/SignatureDTO.java
@@ -1,0 +1,4 @@
+package walab.cap2.dongeui.signature;
+
+public class SignatureDTO {
+}
