+++ src/main/java/com/example/backend/signature/controller/SignatureController.java
@@ -1,0 +1,4 @@
+package walab.cap2.dongeui.signature;
+
+public class SignatureController {
+}
