+++ src/main/java/com/example/backend/signature/repository/SignatureRepository.java
@@ -1,0 +1,6 @@
+package walab.cap2.dongeui.signature;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SignatureRepository extends JpaRepository<Signature, Long> {
+}
